gemini-skill/backend/requirements.txt
modal>=0.66.0
ask-sdk-core>=1.19.0
google-genai>=1.0.0
fastapi>=0.115.0
